requirements.txt
langchain==0.2.5
langchain-community==0.2.5
langgraph==0.0.69
python-dotenv==1.0.1